+++ Firefox/jslib/utils.js
@@ -1,0 +1,44 @@
+function userImport(e){
+    var fileInput = document.createElement('input');
+    fileInput.id = "fileInput"
+    fileInput.type = 'file';
+    fileInput.accept = '.json';
+    fileInput.classList.add('invisible');
+    fileInput.click();
+    fileInput.addEventListener('change', importFromFile);
+}
+
+function importFromFile(e){
+    const fileInput = e.target;
+    const file = e.target.files[0];
+    var reader = new FileReader();
+    reader.onloadend = function() {
+        try {
+            var json = JSON.parse(reader.result);
+            chrome.storage.local.set({['aiUserCommands']: json})
+            .then(() => showMessage('User Commands imported successfully.', 'success'))
+            .catch(e => console.error('>>>', e));
+            aiUserCommands = json;
+        } catch (err) {
+            console.error('>>>', err);
+        } finally{
+            fileInput.remove();
+        }
+    };
+
+    reader.readAsText(file);
+}
+
+async function exportAsFile(e) {
+    let storageKey = 'aiUserCommands';
+    let fileName = `user_commands_export_${(new Date).toISOString().split('T')[0].replace(/\D/g, '')}`;
+
+    const obj = await chrome.storage.local.get([storageKey]);
+    const json = obj[storageKey] || [];
+    var blob = new Blob([JSON.stringify(json, null, 4)], { type: 'application/json' });
+    var url = URL.createObjectURL(blob);
+    var link = document.createElement('a');
+    link.href = url;
+    link.download = `${fileName}.json`;
+    link.click();
+}